split etls, fixed bug on data load
