+++ pyproject.toml
@@ -10,10 +10,13 @@ python = "^3.7"
 lxml = "^4.6.2"
 sqlparse = "^0.4.1"
 pyodbc = "^4.0.30"
+requests = "^2.27.1"
+xmltodict = "^0.12.0"
 
 
 [tool.poetry.dev-dependencies]
 pre-commit = "^2.9.3"
+tox = "^3.24.5"
 
 [tool.black]
 max_line_length = 99
