+++ tox.ini
@@ -3,8 +3,8 @@ envlist = lint
 
 [testenv:reformat]
 commands =
-	black crystal_parser.py main.py settings.py sqlize.py
-	isort crystal_parser.py main.py settings.py sqlize.py
+	black .
+	isort .
 deps =
 	black
 	isort
@@ -37,12 +37,17 @@ deps =
     flake8-markdown
     pygments
 commands =
-    flake8 crystal_parser.py main.py settings.py sqlize.py
-    black --fast --check crystal_parser.py main.py settings.py sqlize.py
-    pylint crystal_parser.py main.py settings.py sqlize.py
+    flake8 .
+    black --fast --check .
+    touch __init__.py
+    pylint .
+    rm __init__.py
 skip_install = true
+whitelist_externals=
+    rm
+    touch
 
 [flake8]
 max-line-length = 99
 exclude = .git,__pycache__,.tox,.venv,venv
-extend-ignore = S608,E203,W503,RST304,E402,T001,S404,S603,S410,S320
+extend-ignore = S608,E203,W503,RST304,E402,T001,S404,S603,S410,S320,E501,SIM113,E800,S105
